useKey hook to perform keydown event
